+++ exportplan/urls.py
@@ -13,5 +13,6 @@ urlpatterns = [
         name='ajax-recommended-countries-data'
     ),
     path('api/export-plan/', views.UpdateExportPlanAPIView.as_view(), name='api-update-export-plan'),
+    path('api/remove-country-data/', ajax.ExportPlanRemoveCountryDataView.as_view(), name='ajax-remove-country-data'),
     path('api/country-data/', ajax.ExportPlanCountryDataView.as_view(), name='ajax-country-data'),
 ]