+++ react-components/src/Services.js
@@ -57,6 +57,10 @@ const getCountryData = function (country) {
   return get(config.countryDataUrl, { country: country }).then((response) => responseHandler(response).json())
 }
 
+const removeCountryData = function (country) {
+  return get(config.removeCountryDataUrl, { country: country }).then((response) => responseHandler(response).json())
+}
+
 const lookupProduct = function ({ q }) {
   return get(config.apiLookupProductUrl, { q }).then((response) => responseHandler(response).json())
 }
@@ -114,6 +118,7 @@ const responseHandler = function (response) {
 let config = {}
 const setConfig = function ({
   countryDataUrl,
+  removeCountryDataUrl,
   countriesBySectorsDataUrl,
   apiLoginUrl,
   apiSignupUrl,
@@ -134,6 +139,7 @@ const setConfig = function ({
   exportPlanTargetMarketsUrl,
 }) {
   config.countryDataUrl = countryDataUrl
+  config.removeCountryDataUrl = removeCountryDataUrl
   config.countriesBySectorsDataUrl = countriesBySectorsDataUrl
   config.apiLoginUrl = apiLoginUrl
   config.apiSignupUrl = apiSignupUrl
@@ -162,6 +168,7 @@ export default {
   store: store,
   updateCompany,
   getCountryData,
+  removeCountryData,
   getCountriesDataBySectors,
   updateExportPlan,
   lookupProduct,