+++ tests/browser/test_export_plan_about_your_plan.py
@@ -1,18 +1,13 @@
-from unittest import mock
-
+# -*- coding: utf-8 -*-
 import pytest
 
-from core import helpers as core_helpers
-from directory_api_client import api_client
-from exportplan import helpers as exportplan_helpers
 from tests.browser.common_selectors import (
     ExportPlanTargetMarkets,
     HeaderCommon,
     HeaderSignedIn,
     StickyHeader,
 )
-from tests.browser.util import should_see_all_elements
-from tests.helpers import create_response
+from tests.browser.steps import should_see_all_expected_page_sections, visit_page
 
 pytestmark = [
     pytest.mark.browser,
@@ -22,75 +17,13 @@ pytestmark = [
 
 
 @pytest.mark.django_db
-@mock.patch.object(exportplan_helpers, 'get_recommended_countries')
-@mock.patch.object(exportplan_helpers, 'update_exportplan')
-@mock.patch.object(exportplan_helpers, 'get_exportplan_marketdata')
-@mock.patch.object(api_client.dataservices, 'get_lastyearimportdata')
-@mock.patch.object(api_client.dataservices, 'get_corruption_perceptions_index')
-@mock.patch.object(api_client.dataservices, 'get_easeofdoingbusiness')
-@mock.patch.object(api_client.exportplan, 'exportplan_list')
-@mock.patch.object(core_helpers, 'get_dashboard_export_opportunities')
-@mock.patch.object(core_helpers, 'get_dashboard_events')
-@mock.patch.object(core_helpers, 'update_company_profile')
-def test_export_plan_about_your_business(
-    mock_update_company_profile,
-    mock_get_dashboard_events,
-    mock_get_dashboard_export_opportunities,
-    mock_get_exportplan,
-    mock_easeofdoingbusiness,
-    mock_cpi,
-    mock_lastyearimportdata,
-    mock_get_exportplan_marketdata,
-    mock_update_exportplan,
-    mock_get_recommended_countries,
-    server_user_browser_dashboard,
+def test_export_plan_about_your_business_page(
+    server_user_browser_dashboard, mock_all_dashboard_and_export_plan_requests_and_responses
 ):
-    mock_update_company_profile.return_value = create_response()
-    mock_get_dashboard_events.return_value = []
-    mock_get_dashboard_export_opportunities.return_value = []
+    live_server, _, browser = server_user_browser_dashboard
 
-    data = [
-        {
-            'export_countries': ['UK'],
-            'export_commodity_codes': [100],
-            'target_markets': [{'country': 'China'}],
-            'rules_regulations': {'country_code': 'CHN'},
-            'sectors': ['Automotive'],
-            'pk': 1,
-        }
-    ]
-    mock_get_exportplan.return_value = create_response(data)
+    visit_page(live_server, browser, 'exportplan:target-markets', 'Target markets')
 
-    mock_get_recommended_countries.return_value = [{'country': 'Japan'}, {'country': 'South Korea'}]
-
-    easeofdoingbusiness_data = {
-        'country_name': 'China',
-        'country_code': 'CHN',
-        'cpi_score_2019': 41,
-        'rank': 80,
-    }
-    mock_easeofdoingbusiness.return_value = create_response(
-        status_code=200, json_body=easeofdoingbusiness_data
+    should_see_all_expected_page_sections(
+        browser, [HeaderCommon, HeaderSignedIn, StickyHeader, ExportPlanTargetMarkets]
     )
-
-    cpi_data = {
-        'country_name': 'China',
-        'country_code': 'CHN',
-        'cpi_score_2019': 41,
-        'rank': 80,
-    }
-    mock_cpi.return_value = create_response(status_code=200, json_body=cpi_data)
-
-    mock_lastyearimportdata.return_value = create_response(
-        status_code=200, json_body={'lastyear_history': 123}
-    )
-
-    mock_get_exportplan_marketdata.return_value = {'timezone': 'Asia/Shanghai', }
-
-    live_server, user, browser = server_user_browser_dashboard
-
-    browser.get(live_server.url + '/export-plan/section/target-markets/')
-    should_see_all_elements(browser, HeaderCommon)
-    should_see_all_elements(browser, HeaderSignedIn)
-    should_see_all_elements(browser, StickyHeader)
-    should_see_all_elements(browser, ExportPlanTargetMarkets)